SQL FOR Pizza Sales Report
In this Project, i used SQL to drive insight from a Pizza Data. The aim of the project is design a dashboard that will enable the management to monitor there sales, profitability and to prepare adequately for Peak period. This Project also helps the management to know the product that needs more attention in terms of marketing and creating awareness.